move encoders into sub-folder
