+++ src/Menu.ts
@@ -1,4 +1,4 @@
-import BinaryEncoder from "./BinaryEncoder.ts";
+import BinaryEncoder from "./encoding/BinaryEncoder.ts";
 
 export type EncodingType = "Fixed" | "Variable" | "Other";
